backend_code/requirements.txt
Django
djangorestframework
psycopg2
selenium
webdriver-manager
PILLOW
pytesseract
django-cors-headers
django-cors-headers
googletrans-3.1.0a0
gTTS

